+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+2.0.6
+-----
+#. Make object detail template resolution follow the standard Django naming conventions. Backward compatibility is preserved.
+#. Move test templates into tests directory.
+#. Add dependency on ``django-ultracache`` and defer ``jmbocache`` template tag to ``ultracache`` template tag.
+
 2.0.5
 -----
 #. Gracefully handle missing images in the API.